// cloudflare/src/request.generated.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Evidence = string;
/**
 * @minItems 1
 * @maxItems 11
 */
export type EventIds = string[];
export type Reviewed = true;
export type IdempotencyKey = string;

export interface AnalysisRequest {
  evidence: Evidence;
  event_ids: EventIds;
  reviewed: Reviewed;
  idempotency_key: IdempotencyKey;
}